# backend/docs/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminCountTable" : {
        "properties" : {
          "columns" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "rows" : {
            "items" : {
              "items" : { },
              "type" : "array"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "AdminReportStats" : {
        "properties" : {
          "byReason" : {
            "$ref" : "#/components/schemas/AdminCountTable"
          },
          "byStatus" : {
            "$ref" : "#/components/schemas/AdminCountTable"
          },
          "daily" : {
            "$ref" : "#/components/schemas/AdminCountTable"
          },
          "days" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "HideResult" : {
        "properties" : {
          "affected" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "ListResponse" : {
        "properties" : {
          "reports" : {
            "items" : {
              "$ref" : "#/components/schemas/ReportDetail"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ReportDetail" : {
        "properties" : {
          "contextKey" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "memo" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReportedUserListResponse" : {
        "properties" : {
          "users" : {
            "items" : {
              "$ref" : "#/components/schemas/ReportedUserSummary"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ReportedUserSummary" : {
        "properties" : {
          "fromDeletedAccounts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "lastReportedAt" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "reasons" : {
            "additionalProperties" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "object"
          },
          "reportCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "reporterCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "suspended" : {
            "type" : "boolean"
          },
          "userId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ReviewReportRequest" : {
        "properties" : {
          "decision" : {
            "type" : "boolean"
          },
          "status" : {
            "enum" : [
              "PENDING",
              "ACTIONED",
              "DISMISSED"
            ],
            "type" : "string"
          }
        },
        "required" : [
          "status"
        ],
        "type" : "object"
      },
      "ReviewResult" : {
        "properties" : {
          "reviewed" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "UpdatePresenceRequest" : {
        "properties" : {
          "sessionId" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "sessionKind" : {
            "enum" : [
              "LOBBY",
              "MATCH"
            ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "기기 식별자 기반 계정, 유저 검색, 친구, 차단, 접속 상태 API.\n\nX-User-Id 는 인증이 아니라 식별입니다. 되돌릴 수 없는 연산만 자격증명(X-Device-Id)을 함께 요구합니다. 오류 코드 표와 하트비트 규칙은 docs/client-guide.md 를 보세요.",
    "title" : "D205 API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/reports" : {
      "get" : {
        "operationId" : "list_3",
        "parameters" : [
          {
            "in" : "query",
            "name" : "status",
            "required" : false,
            "schema" : {
              "default" : "PENDING",
              "enum" : [
                "PENDING",
                "ACTIONED",
                "DISMISSED"
              ],
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReportedUserListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
    "/api/v1/admin/reports/entries/{reportId}" : {
      "delete" : {
        "operationId" : "purgeEntry",
        "parameters" : [
          {
            "in" : "path",
            "name" : "reportId",
            "required" : true,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HideResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
    "/api/v1/admin/reports/entries/{reportId}/hidden" : {
      "patch" : {
        "operationId" : "hideEntry",
        "parameters" : [
          {
            "in" : "path",
            "name" : "reportId",
            "required" : true,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HideResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
...
    "/api/v1/admin/reports/stats" : {
      "get" : {
        "operationId" : "stats",
        "parameters" : [
          {
            "in" : "query",
            "name" : "days",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminReportStats"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
    "/api/v1/admin/reports/{userId}" : {
      "delete" : {
        "operationId" : "purge",
        "parameters" : [
          {
            "in" : "path",
            "name" : "userId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "status",
            "required" : false,
            "schema" : {
              "enum" : [
                "PENDING",
                "ACTIONED",
                "DISMISSED"
              ],
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HideResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      },
      "get" : {
        "operationId" : "detail",
        "parameters" : [
          {
            "in" : "path",
            "name" : "userId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "status",
            "required" : false,
            "schema" : {
              "enum" : [
                "PENDING",
                "ACTIONED",
                "DISMISSED"
              ],
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      },
      "patch" : {
        "operationId" : "review",
        "parameters" : [
          {
            "in" : "path",
            "name" : "userId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewReportRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReviewResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
    "/api/v1/admin/reports/{userId}/hidden" : {
      "patch" : {
        "operationId" : "hide",
        "parameters" : [
          {
            "in" : "path",
            "name" : "userId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "status",
            "required" : false,
            "schema" : {
              "enum" : [
                "PENDING",
                "ACTIONED",
                "DISMISSED"
              ],
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HideResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "admin-report-controller"
        ]
      }
    },
...
    "/api/v1/presence" : {
      "delete" : {
        "operationId" : "goOffline",
        "parameters" : [
          {
            "in" : "header",
            "name" : "X-User-Id",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [
          "presence-controller"
        ]
      },
      "put" : {
        "operationId" : "report",
        "parameters" : [
          {
            "in" : "header",
            "name" : "X-User-Id",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdatePresenceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [
          "presence-controller"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "description" : "운영",
      "url" : "https://j15d205.p.ssafy.io"
    },
    {
      "description" : "로컬",
      "url" : "http://localhost:8080"
    }
  ]
}